Added exceptions to Movie

+++ ticket-service/src/main/java/com/epam/training/ticketservice/repository/impl/JpaMovieRepository.java
@@ -3,9 +3,12 @@ package com.epam.training.ticketservice.repository.impl;
 import com.epam.training.ticketservice.dataaccess.dao.MovieDao;
 import com.epam.training.ticketservice.dataaccess.entity.MovieEntity;
 import com.epam.training.ticketservice.repository.MovieRepository;
+import com.epam.training.ticketservice.service.ServiceException.MovieAlreadyExistsException;
+import com.epam.training.ticketservice.service.ServiceException.MovieNotFoundException;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public class JpaMovieRepository implements MovieRepository {
@@ -17,7 +20,10 @@ public class JpaMovieRepository implements MovieRepository {
     }
 
     @Override
-    public void createMovie(MovieEntity movie) {
+    public void createMovie(MovieEntity movie) throws MovieAlreadyExistsException {
+        if(isMovieExists(movie.getTitle())) {
+            throw new MovieAlreadyExistsException(("Movie already exists"));
+        }
         MovieEntity newMovie = new MovieEntity(movie.getTitle(), movie.getGenre(), movie.getRuntime());
         movieDao.save(newMovie);
     }
@@ -29,18 +35,32 @@ public class JpaMovieRepository implements MovieRepository {
     }
 
     @Override
-    public void updateMovie(String title, String genre, int runtime) {
-        MovieEntity movieEntity = movieDao.findMovieByTitle(title);
-        movieEntity.setTitle(title);
-        movieEntity.setGenre(genre);
-        movieEntity.setRuntime(runtime);
-        movieDao.save(movieEntity);
+    public void updateMovie(String title, String genre, int runtime) throws MovieNotFoundException {
+        if (isMovieExists(title)) {
+            MovieEntity movieEntity = movieDao.findMovieByTitle(title);
+            movieEntity.setTitle(title);
+            movieEntity.setGenre(genre);
+            movieEntity.setRuntime(runtime);
+            movieDao.save(movieEntity);
+        } else {
+            throw new MovieNotFoundException("Movie not found");
+        }
+
     }
 
     @Override
-    public void deleteMovie(String title) {
-        MovieEntity movieEntity = movieDao.findMovieByTitle(title);
-        movieDao.delete(movieEntity);
+    public void deleteMovie(String title) throws MovieNotFoundException {
+        if (isMovieExists(title)) {
+            MovieEntity movieEntity = movieDao.findMovieByTitle(title);
+            movieDao.delete(movieEntity);
+        } else {
+            throw new MovieNotFoundException("Movie not found");
+        }
+    }
+
+    private boolean isMovieExists(String title) {
+        Optional<MovieEntity> movieEntity = Optional.ofNullable(movieDao.findMovieByTitle(title));
+        return movieEntity.isPresent();
     }
 
 }
